Ignore is-installed if startup script is disabled
If the Seattle installer is told to not add the startup script
anyway, then why check if Seattle is currently installed?

In order to ignore an existing installation on `--disable-startup-script`,
we parse the command-line flags *before* doing the is-installed check.
This lets us perform the check only if we should enable the startup
script.

(Part of why this is relevant is how `popen` currently fails on
Android and must be replaced with a custom method of ours. However,
we cannot use `crond` on Android anyway, so we disable the startup
script on installation, and it makes sense to skip the is-installed
check then, too.)